Clean webGL.JS code snippet
