+++ src/utils/checkWebGL.ts
@@ -1,0 +1,24 @@
+/* Base taken from:
+ * https://github.com/mrdoob/three.js/blob/master/examples/jsm/WebGL.js */
+
+export const isWebGLAvailable = () => {
+    try {
+        const canvas = document.createElement('canvas')
+        return !!(
+            window.WebGLRenderingContext &&
+            (canvas.getContext('webgl') ||
+                canvas.getContext('experimental-webgl'))
+        )
+    } catch (e) {
+        return false
+    }
+}
+
+export const isWebGL2Available = () => {
+    try {
+        const canvas = document.createElement('canvas')
+        return !!(window.WebGL2RenderingContext && canvas.getContext('webgl2'))
+    } catch (e) {
+        return false
+    }
+}
